✨ add functionality to reactivate categories and products

+++ app/view/components/produtosCards.php
@@ -9,7 +9,6 @@ $paginaAtual = basename($_SERVER['SCRIPT_NAME']);
         
         
         <p class="cards-titulo w-75 text-center" title="<?= htmlspecialchars($produto->getNome()) ?>"><?= htmlspecialchars($produto->getNome()) ?></p>
-        <p class="cards-titulo w-75 text-center" title="<?= htmlspecialchars($produto->getId()) ?>"><?= htmlspecialchars($produto->getId()) ?></p>
         
         <span>R$ <?= htmlspecialchars($produto->getPreco()) ?></span>
     </div>
@@ -19,9 +18,16 @@ $paginaAtual = basename($_SERVER['SCRIPT_NAME']);
             <?php 
                 $redirectToExcluir = 'excluirSabor.php?idProduto=' . urlencode($produto->getId()) . '&idCategoria=' . urlencode($produto->getCategoria());
                 $redirectToEditar  = 'editarSabor.php?idProduto=' . urlencode($produto->getId()) . '&idCategoria=' . urlencode($produto->getCategoria());
+                $redirectToReativar = 'gerenciarProdutos.php?produtoAtivar=' . urlencode($produto->getId()) . '&idCategoria=' . urlencode($produto->getCategoria());
             ?>
-                <a class="botao botao-primary" href="<?= $redirectToEditar ?>">Editar</a>
-                <a class="botao botao-alerta" href="<?= $redirectToExcluir ?>">Excluir</a>
+
+                <?php if ($produto->getDesativado() != 1): ?>
+                    <a class="botao botao-primary" href="<?= $redirectToEditar ?>">Editar</a>
+                    <a class="botao botao-alerta" href="<?= $redirectToExcluir ?>">Excluir</a>
+
+                <?php else: ?>
+                    <a class="botao botao-primary" href="<?= htmlspecialchars($redirectToReativar) ?>">Reativar</a>
+                <?php endif; ?>
         <?php else: ?>
             <a class="botao botao-primary" href="carrinho.php?add=<?= urlencode($produto->getId()) ?>">Adicionar ao Carrinho</a>
         <?php endif; ?>
